Day 3

+++ backend/src/main/java/com/wecp/progressive/entity/Product.java
@@ -6,11 +6,14 @@ private int warehouseId;
 private String productName;
 private String productDescription;
 private int quantity;
-private double price;
+// private double price;
+private Long price;
+
 public Product() {
 }
+
 public Product(int productId, int warehouseId, String productName, String productDescription, int quantity,
-        double price) {
+Long price) {
     this.productId = productId;
     this.warehouseId = warehouseId;
     this.productName = productName;
@@ -48,10 +51,10 @@ public int getQuantity() {
 public void setQuantity(int quantity) {
     this.quantity = quantity;
 }
-public double getPrice() {
+public Long getPrice() {
     return price;
 }
-public void setPrice(double price) {
+public void setPrice(Long price) {
     this.price = price;
 }
